requirements.txt
asgiref==3.11.1
dj-database-url==3.1.2
Django==5.2.12
django-heroku==0.3.1
djangorestframework==3.16.1
drf-yasg==1.21.15
gunicorn==25.1.0
inflection==0.5.1
packaging==26.0
psycopg2==2.9.11
pytz==2026.1.post1
PyYAML==6.0.3
sqlparse==0.5.5
uritemplate==4.2.0
whitenoise==6.12.0
